test(tui): split gateway chat coverage
